bootstrap: Fix LLVM bin path setup for Windows.
